Fix run command hanging for secure protocol device
Problem

The `flutter-tizen run` command roughly takes the following steps:
- build tpk
- stop the app on target
- install the app on target
- start the app on target

For secure_protocol enabled devices, `sdb shell 0 kill <app_name>`
is executed to stop the app. At this point, if the app is running on the device,
the command will kill the app, otherwise it will report that the app
is already terminated. However, if the app was not even installed
in the first place(applist doesn't list the app), the command is irresponsive.

Solution

A quick solution is checking if the app is installed on the device before
running the command to stop the app. This adds about 500ms to the
`flutter-tizen run` command.