Reduce build time by using cache files
The FlutterBuildSystem supports caching intermediate build files
of each target to reduce the overall build time.
In order to use this functionality, we need to specify the files
that each target depends on. If those files are always known during the
dev-lifecycle, they can be listed in Target.inputs or Target.outputs.
In most cases, however, new dependencies such as plugins or source
code files will be added during development. To handle these cases,
we can either (1) use wildcard patterns in Source.pattern(), or (2)
create a DepFile that stores the dependencies of a Target at runtime
and specify the DepFile filename in Target.depfiles.

Although (1) is a much simpler solution, the implementation of resolving
the wildcard pattern doesn't handle our use cases, such as:
- doesn't support recursive file search
- throws error when intermediate directories are missing in the file path
We need at least one of the above functionalities because our intermediate
files are layed out in a multiple depth folder hierarchy.

As such, I've choosen to use solution (2) despite increasing the code
complexity of maintaining depfiles. It is crucial for depFiles to list
all input and output files that participated in Target.build().