add linker option "-Wl,-E" to remove regression
When the toolchain is changed, the linker option "-Wl,-E" is omitted.
Due to this, performance degradation occurred by launching the app using exec() in the candidate process.

Add "-Wl,-E" to linker option of cpp template so that main() symbol are added to the dynamic symbol table,
and can be used in candidate process.