Change ui thread policy to SeparateThread
This commit reverts part of #678.
The default value of UI thread policy on other platforms(such as Linux) is PlatformThread(merged thread).
Since most examples(plugins) on Tizen worked without any issues, we applied the same policy.
However, an issue was discovered in `tizen_interop_callbacks`.
A thread error occurred even when calling blocking APIs from an isolate.
While improvements are needed for the tizen_interop_callbacks package,
we will first apply a hotfix to revert to the previous UI thread policy.
This patch may be reverted once the issue is resolved.